pyproject.toml
[project]
name = "fastapi-agent-chat"
version = "0.1.0"
description = "FastAPI agent with immediate stop functionality using Redis signals"
readme = "README.md"
requires-python = "==3.13.*"
license = {text = "MIT"}
authors = [
    {name = "Developer", email = "dev@example.com"}
]

dependencies = [
    # FastAPI and ASGI
    "fastapi[standard]>=0.110.0",
    # Redis for signal bus
    "redis[hiredis]>=5.0.0",
    "asyncio-redis>=0.9.0",
    # Core utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "python-multipart>=0.0.9",
    # Logging
    "loguru>=0.7.0",
    # Strands AI Agents
    "strands-agents>=1.14.0",
    "strands-agents-tools>=0.2.13",
    # AWS SDK for Bedrock
    "boto3>=1.35.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
# Set line length
line-length = 100
indent-width = 4

# Target Python version
target-version = "py313"

# Exclude directories
exclude = [
    ".git",
    ".venv",
    ".ruff_cache",
    ".mypy_cache",
    "__pycache__",
    "build",
    "dist",
    "logs",
    "data",
]

[tool.ruff.lint]
# Enable rule sets
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "B",     # flake8-bugbear
    "I",     # isort
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "C4",    # flake8-comprehensions
    "RUF",   # Ruff-specific rules
]

# Disable specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Function calls in argument defaults
]

# Allow automatic fixes
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.ruff.format]
# Use double quotes
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Respect magic trailing commas
skip-magic-trailing-comma = false

# Auto-detect line ending
line-ending = "auto"

# Format docstrings
docstring-code-format = true

[tool.mypy]
# Python version
python_version = "3.13"

# Strictness
strict = false
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = false
strict_equality = true
check_untyped_defs = true

# Import behavior
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true

# Output
show_error_codes = true
show_column_numbers = true
pretty = true
color_output = true

# Caching
cache_dir = ".mypy_cache"
sqlite_cache = true

# Exclude patterns
exclude = [
    "^build/",
    "^dist/",
    "^logs/",
    "^data/",
]
